+++ examples/minimal/variables.tf
@@ -27,3 +27,13 @@ variable "egress_cidr_blocks" {
   type = list(string)
   default = ["0.0.0.0/0"]
 }
+
+variable "tags" {
+  type        = map(string)
+  description = "A map of tags to add to all resources created by this example."
+  default     = { 
+    Author = "Tamr"
+    Environment = "Example"
+  }
+}
+